requirements.txt
torch>=2.0.0
transformers>=4.36.0
safetensors>=0.4.0
huggingface_hub>=0.20.0
gradio==4.44.1
pypdf>=4.0.0
python-docx>=1.0.0